+++ src-core/src/main.rs
@@ -84,6 +84,8 @@ fn configure_command_handlers() -> impl Fn(tauri::Invoke) {
         os::commands::is_semver_higher,
         os::commands::upgrade_net_core,
         os::commands::upgrade_asp_net_core,
+        os::commands::set_windows_power_policy,
+        os::commands::active_windows_power_policy,
         osc::commands::osc_send_bool,
         osc::commands::osc_send_float,
         osc::commands::osc_send_int,