+++ src-ui/app/views/dashboard-view/views/power-automations-view/power-automations-view.component.ts
@@ -8,5 +8,7 @@ import { noop } from '../../../../utils/animations';
   animations: [noop()],
 })
 export class PowerAutomationsViewComponent {
-  activeTab: 'CONTROLLERS_AND_TRACKERS' | 'BASE_STATIONS' = 'CONTROLLERS_AND_TRACKERS';
+  // activeTab: 'CONTROLLERS_AND_TRACKERS' | 'BASE_STATIONS' | 'WINDOWS_POWER_POLICY' = 'CONTROLLERS_AND_TRACKERS';
+  activeTab: 'CONTROLLERS_AND_TRACKERS' | 'BASE_STATIONS' | 'WINDOWS_POWER_POLICY' =
+    'WINDOWS_POWER_POLICY';
 }