+++ src-ui/app/views/dashboard-view/views/sleep-detection-view/battery-percentage-enable-sleepmode-modal/battery-percentage-enable-sleep-mode-modal.component.ts
@@ -55,6 +55,5 @@ export class BatteryPercentageEnableSleepModeModalComponent
     } else {
       this.triggerClasses.push(deviceClass);
     }
-
   }
 }